chore(notifications): move shares styles to common/Notifications/
