+++ packages/components/src/Banner/Banner.tsx
@@ -6,7 +6,7 @@ import CloseButton from "../common/CloseButton";
 import React from "react";
 import Text from "../Text";
 import clsx from "clsx";
-import colorStyles from "../common/Notification.module.css";
+import colorStyles from "../common/Notifications/Notification.module.css";
 import styles from "./Banner.module.css";
 
 type Props = {
